Switch from minitrace to Tracy 0.7

+++ SanityEngine/extern/rex/include/rx/core/time/delay.cpp
@@ -28,7 +28,7 @@ void delay(Uint64 _milliseconds) {
     was_error = nanosleep(&tv, &elapsed);
   } while (was_error && (errno == EINTR));
 #elif defined(RX_PLATFORM_WINDOWS)
-  Sleep(_milliseconds);
+  Sleep(static_cast<DWORD>(_milliseconds));
 #endif
 }
